requirements.txt
black==22.3.0
Django==4.0.3
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
Markdown==3.5.1
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0